refactor: split production, diplomacy, and leader-dialog part files (#3878)
Extract production Available subpanel section builders, diplomacy panel
constants/filter assembly, and leader-selection responsive picker chrome
into library parts to shrink remaining monolith hosts.

Refs #3878
